requirements/local.txt
-r base.txt #includes the base.txt  requirement file
coverage==5.1
django-debug-toolbar==4.0.0
djangorestframework==3.14.0

